+++ src/plugins/data_source_management/public/components/data_source_aggregated_view/data_source_aggregated_view.test.tsx
@@ -8,22 +8,27 @@ import React from 'react';
 import { DataSourceAggregatedView } from './data_source_aggregated_view';
 import { SavedObjectsClientContract } from '../../../../../core/public';
 import { notificationServiceMock } from '../../../../../core/public/mocks';
+import { mockManagementPlugin } from '../../mocks';
 import { getDataSourcesWithFieldsResponse, mockResponseForSavedObjectsCalls } from '../../mocks';
-import { render } from '@testing-library/react';
 
 describe('DataSourceAggregatedView', () => {
   let component: ShallowWrapper<any, Readonly<{}>, React.Component<{}, {}, any>>;
   let client: SavedObjectsClientContract;
   const { toasts } = notificationServiceMock.createStartContract();
+  const mockedContext = mockManagementPlugin.createDataSourceManagementContext();
+  const uiSettings = mockedContext.uiSettings;
 
   beforeEach(() => {
+    jest.clearAllMocks();
+
     client = {
       find: jest.fn().mockResolvedValue([]),
     } as any;
     mockResponseForSavedObjectsCalls(client, 'find', getDataSourcesWithFieldsResponse);
   });
 
-  it('should render normally with local cluster not hidden and all options', () => {
+  it('should render normally with local cluster not hidden and all options', async () => {
+    spyOn(uiSettings, 'get').and.returnValue('test1');
     component = shallow(
       <DataSourceAggregatedView
         fullWidth={false}
@@ -31,14 +36,19 @@ describe('DataSourceAggregatedView', () => {
         savedObjectsClient={client}
         notifications={toasts}
         displayAllCompatibleDataSources={true}
+        uiSettings={uiSettings}
       />
     );
+    await component.instance().componentDidMount!();
     expect(component).toMatchSnapshot();
+    expect(uiSettings.get).toBeCalledWith('defaultDataSource', null);
+    expect(component.state('defaultDataSource')).toBe('test1');
     expect(client.find).toBeCalledWith({
       fields: ['id', 'title', 'auth.type'],
       perPage: 10000,
       type: 'data-source',
     });
+    expect(component.state('dataSourceOptions')).toHaveLength(4);
     expect(toasts.addWarning).toBeCalledTimes(0);
   });
 
@@ -60,8 +70,6 @@ describe('DataSourceAggregatedView', () => {
       type: 'data-source',
     });
     expect(toasts.addWarning).toBeCalledTimes(0);
-    const badge = container.find('EuiNotificationBadge').text();
-    expect(badge).toEqual('0');
   });
 
   it('should render normally with local cluster and actice selections', () => {
@@ -82,8 +90,6 @@ describe('DataSourceAggregatedView', () => {
       type: 'data-source',
     });
     expect(toasts.addWarning).toBeCalledTimes(0);
-    const badge = container.find('EuiNotificationBadge').text();
-    expect(badge).toEqual('1');
   });
 
   it('should render normally with data source filter', () => {
@@ -104,23 +110,5 @@ describe('DataSourceAggregatedView', () => {
       type: 'data-source',
     });
     expect(toasts.addWarning).toBeCalledTimes(0);
-    const badge = container.find('EuiNotificationBadge').text();
-    expect(badge).toEqual('All');
-  });
-
-  it('should render popup when clicking on info icon', async () => {
-    const container = render(
-      <DataSourceAggregatedView
-        fullWidth={false}
-        hideLocalCluster={false}
-        savedObjectsClient={client}
-        notifications={toasts}
-        displayAllCompatibleDataSources={false}
-        activeDataSourceIds={['test1']}
-      />
-    );
-    const infoIcon = await container.findByTestId('dataSourceAggregatedViewInfoButton');
-    infoIcon.click();
-    expect(container).toMatchSnapshot();
   });
 });